chore: migrate to src/ dir for next and vite examples, move wagmi setup to wagmi.ts
