+++ examples/with-create-react-app/src/index.tsx
@@ -2,32 +2,13 @@ import '@rainbow-me/rainbowkit/styles.css';
 import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { getDefaultConfig, RainbowKitProvider } from '@rainbow-me/rainbowkit';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { WagmiProvider } from 'wagmi';
-import {
-  arbitrum,
-  base,
-  mainnet,
-  optimism,
-  polygon,
-  sepolia,
-} from 'wagmi/chains';
+import { RainbowKitProvider } from '@rainbow-me/rainbowkit';
 
 import App from './App';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-const config = getDefaultConfig({
-  appName: 'RainbowKit demo',
-  projectId: 'YOUR_PROJECT_ID',
-  chains: [
-    mainnet,
-    polygon,
-    optimism,
-    arbitrum,
-    base,
-    ...(process.env.REACT_APP_ENABLE_TESTNETS === 'true' ? [sepolia] : []),
-  ],
-});
+import { config } from './wagmi';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
