+++ examples/with-next-app-i18n/src/app/[locale]/page.tsx
