+++ examples/with-next-app-i18n/src/app/[locale]/providers.tsx
@@ -1,0 +1,29 @@
+'use client';
+
+import * as React from 'react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { WagmiProvider } from 'wagmi';
+import {
+  RainbowKitProvider,
+  type Locale,
+} from '@rainbow-me/rainbowkit';
+
+import { config } from '../../wagmi';
+
+const queryClient = new QueryClient();
+
+export function Providers({
+  children,
+  locale,
+}: {
+  children: React.ReactNode;
+  locale: Locale;
+}) {
+  return (
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>
+        <RainbowKitProvider locale={locale}>{children}</RainbowKitProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
+  );
+}
