+++ examples/with-next-app-i18n/src/app/layout.tsx
