+++ examples/with-next-app/src/app/layout.tsx
