+++ examples/with-next-app/src/app/page.tsx
