+++ examples/with-next-custom-button/src/pages/index.tsx
