+++ examples/with-next-rainbow-button/src/pages/_app.tsx
@@ -1,0 +1,25 @@
+import '../styles/global.css';
+import '@rainbow-me/rainbow-button/styles.css';
+import type { AppProps } from 'next/app';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { WagmiProvider } from 'wagmi';
+import { RainbowButtonProvider } from '@rainbow-me/rainbow-button';
+
+import { config } from '../wagmi';
+
+const queryClient = new QueryClient();
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>
+        <RainbowButtonProvider>
+          <Component {...pageProps} />
+        </RainbowButtonProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
+  );
+}
+
+export default MyApp;
