+++ examples/with-next-rainbow-button/src/pages/index.tsx
