--- examples/with-next-siwe-iron-session/pages/api/logout.ts
@@ -1,18 +1,0 @@
-import { withIronSessionApiRoute } from 'iron-session/next';
-import { NextApiRequest, NextApiResponse } from 'next';
-import { ironOptions } from './../../lib/iron';
-
-const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const { method } = req;
-  switch (method) {
-    case 'GET':
-      req.session.destroy();
-      res.send({ ok: true });
-      break;
-    default:
-      res.setHeader('Allow', ['GET']);
-      res.status(405).end(`Method ${method} Not Allowed`);
-  }
-};
-
-export default withIronSessionApiRoute(handler, ironOptions);
