+++ examples/with-next-siwe-iron-session/src/pages/_app.tsx
@@ -1,0 +1,121 @@
+// This example is based on the wagmi SIWE tutorial
+// https://wagmi.sh/examples/sign-in-with-ethereum
+import '../styles/global.css';
+import '@rainbow-me/rainbowkit/styles.css';
+import type { AppProps } from 'next/app';
+import { useEffect, useMemo, useRef, useState } from 'react';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { WagmiProvider } from 'wagmi';
+import {
+  RainbowKitProvider,
+  createAuthenticationAdapter,
+  RainbowKitAuthenticationProvider,
+  AuthenticationStatus,
+} from '@rainbow-me/rainbowkit';
+import { SiweMessage } from 'siwe';
+
+const queryClient = new QueryClient();
+
+export default function App({ Component, pageProps }: AppProps) {
+  const fetchingStatusRef = useRef(false);
+  const verifyingRef = useRef(false);
+  const [authStatus, setAuthStatus] = useState<AuthenticationStatus>('loading');
+
+  // Fetch user when:
+  useEffect(() => {
+    const fetchStatus = async () => {
+      if (fetchingStatusRef.current || verifyingRef.current) {
+        return;
+      }
+
+      fetchingStatusRef.current = true;
+
+      try {
+        const response = await fetch('/api/me');
+        const json = await response.json();
+        setAuthStatus(json.address ? 'authenticated' : 'unauthenticated');
+      } catch (_error) {
+        setAuthStatus('unauthenticated');
+      } finally {
+        fetchingStatusRef.current = false;
+      }
+    };
+
+    // 1. page loads
+    fetchStatus();
+
+    // 2. window is focused (in case user logs out of another window)
+    window.addEventListener('focus', fetchStatus);
+    return () => window.removeEventListener('focus', fetchStatus);
+  }, []);
+
+  const authAdapter = useMemo(() => {
+    return createAuthenticationAdapter({
+      getNonce: async () => {
+        const response = await fetch('/api/nonce');
+        return await response.text();
+      },
+
+      createMessage: ({ nonce, address, chainId }) => {
+        return new SiweMessage({
+          domain: window.location.host,
+          address,
+          statement: 'Sign in with Ethereum to the app.',
+          uri: window.location.origin,
+          version: '1',
+          chainId,
+          nonce,
+        });
+      },
+
+      getMessageBody: ({ message }) => {
+        return message.prepareMessage();
+      },
+
+      verify: async ({ message, signature }) => {
+        verifyingRef.current = true;
+
+        try {
+          const response = await fetch('/api/verify', {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify({ message, signature }),
+          });
+
+          const authenticated = Boolean(response.ok);
+
+          if (authenticated) {
+            setAuthStatus(authenticated ? 'authenticated' : 'unauthenticated');
+          }
+
+          return authenticated;
+        } catch (error) {
+          return false;
+        } finally {
+          verifyingRef.current = false;
+        }
+      },
+
+      signOut: async () => {
+        setAuthStatus('unauthenticated');
+        await fetch('/api/logout');
+      },
+    });
+  }, []);
+
+  return (
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>
+        <RainbowKitAuthenticationProvider
+          adapter={authAdapter}
+          status={authStatus}
+        >
+          <RainbowKitProvider>
+            <Component {...pageProps} />
+          </RainbowKitProvider>
+        </RainbowKitAuthenticationProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
+  );
+}
