+++ examples/with-next-siwe-iron-session/src/pages/api/nonce.ts
@@ -1,0 +1,21 @@
+import { withIronSessionApiRoute } from 'iron-session/next';
+import { NextApiRequest, NextApiResponse } from 'next';
+import { generateNonce } from 'siwe';
+import { ironOptions } from '../../../lib/iron';
+
+const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const { method } = req;
+  switch (method) {
+    case 'GET':
+      req.session.nonce = generateNonce();
+      await req.session.save();
+      res.setHeader('Content-Type', 'text/plain');
+      res.send(req.session.nonce);
+      break;
+    default:
+      res.setHeader('Allow', ['GET']);
+      res.status(405).end(`Method ${method} Not Allowed`);
+  }
+};
+
+export default withIronSessionApiRoute(handler, ironOptions);
