+++ examples/with-next-siwe-iron-session/src/pages/api/verify.ts
@@ -1,0 +1,35 @@
+import { withIronSessionApiRoute } from 'iron-session/next';
+import { NextApiRequest, NextApiResponse } from 'next';
+import { SiweMessage } from 'siwe';
+import { ironOptions } from '../../../lib/iron';
+
+const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const { method } = req;
+  switch (method) {
+    case 'POST':
+      try {
+        const { message, signature } = req.body;
+        const siweMessage = new SiweMessage(message);
+        const { success, error, data } = await siweMessage.verify({
+          signature,
+        });
+
+        if (!success) throw error;
+
+        if (data.nonce !== req.session.nonce)
+          return res.status(422).json({ message: 'Invalid nonce.' });
+
+        req.session.siwe = data;
+        await req.session.save();
+        res.json({ ok: true });
+      } catch (_error) {
+        res.json({ ok: false });
+      }
+      break;
+    default:
+      res.setHeader('Allow', ['POST']);
+      res.status(405).end(`Method ${method} Not Allowed`);
+  }
+};
+
+export default withIronSessionApiRoute(handler, ironOptions);
