+++ examples/with-next-siwe-iron-session/src/pages/index.tsx
