+++ examples/with-next-siwe-next-auth/src/pages/api/auth/[...nextauth].ts
