+++ examples/with-next-siwe-next-auth/src/pages/index.tsx
