+++ examples/with-next-wallet-button/src/pages/index.tsx
