+++ examples/with-next/src/pages/index.tsx
