+++ examples/with-vite/src/wagmi.ts
@@ -1,0 +1,9 @@
+
+import { getDefaultConfig } from '@rainbow-me/rainbowkit';
+import { arbitrum, base, mainnet, optimism, polygon } from 'wagmi/chains';
+
+export const config = getDefaultConfig({
+  appName: 'RainbowKit demo',
+  projectId: 'YOUR_PROJECT_ID',
+  chains: [mainnet, polygon, optimism, arbitrum, base],
+});
